+++ include/ingress.h
@@ -4,6 +4,7 @@
 #include "rpc_mapper.h"
 #include "rpc_message.h"
 #include "rpc_queue.h"
+#include <cstddef>
 #include <cstdint>
 #include <deque>
 
@@ -15,12 +16,12 @@ class Ingress {
 
         void enqueue(RPCMessage* rpc);
         RPCMessage* dequeue();
-        int size() const { return queue.size(); }
+        size_t size() const { return queue.size(); }
         void update_p95(int64_t p95);
         bool check_drop(RPCQueue&, RPCMapper&);
 
     private:
         std::deque<RPCMessage*> queue;
         int64_t p95;
-        uint32_t drop_id;
+        int32_t drop_id;
 };
